adds all fields to dummy data, creates HubTable, cleanup
